made a scene that we can use by making the prefad for enviroment...
making a timer on the spawn wave so 2 do not spawn

+++ Assets/Scripts/Enemy/EnemyHolderManager.cs
@@ -9,7 +9,7 @@ public class EnemyHolderManager : MonoBehaviour
     private EnemySpawnerBehavior[] enemySpawnerBehavior;
 
     [SerializeField]
-    private int _timeToSpawnWaves;
+    private float _timeToSpawnWaves;
     private void Awake() { Instace = this; }
     private void Start()
     {
